Refactor: modularized game logic + real-time UI update on game cancellation
- Refactored and reorganized game-related code into smaller, modular functions
- Created new utility files to improve code structure and readability
- Implemented real-time removal of game button from dashboard when a game is canceled
- Set up groundwork for future game cancellation notification
